fix(basicmicro): read command 90 status as 32-bit with a legacy 16-bit retry (PR #729 review)
The manual leaves the status width unstated ('[Status, CRC]'), but current
MCP firmware returns a 32-bit word — the official Basicmicro Arduino
library reads it with Read4. read_status() now reads 4 bytes first; if
that transaction fails (older firmware replying 16-bit ends the reply
mid-read), the receive timeout has already spanned the controller's 10 ms
packet-clear window, so a single 16-bit retry is performed for legacy
units. Status out-param and mask enum widened to uint32_t (the manual's
documented masks occupy the low 16 bits); example updated.

Host tests ALL PASSED; esp32 example builds; cppcheck clean.

+++ components/basicmicro/include/detail/basicmicro_core.hpp
@@ -127,8 +127,10 @@ enum class BasicmicroCommand : uint8_t {
 };
 
 /// @brief Unit status bit masks returned by BasicmicroCommand::ReadStatus
-///        (manual command 90).
-enum class BasicmicroStatus : uint16_t {
+///        (manual command 90). Current MCP firmware returns a 32-bit status
+///        word (read via Read4 in the official library); the manual documents
+///        the masks below, which occupy the low 16 bits.
+enum class BasicmicroStatus : uint32_t {
   Normal = 0x0000,
   M1OverCurrentWarning = 0x0001,
   M2OverCurrentWarning = 0x0002,
